requirements/base.txt
Django==6.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.2
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
channels==4.0.0
channels-redis==4.2.0
pgvector==0.2.5
python-dotenv==1.0.1
cryptography==42.0.5
django-redis==5.4.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
